% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/B_ij_fun_old.R
\name{B_ij_fun_old}
\alias{B_ij_fun_old}
\title{Calculates B_ij: the coincidence probability due to dispersal}
\usage{
B_ij_fun_old(
  graph,
  field_B = "length",
  dir_distance_type = "symmetric",
  disp_type = "exponential",
  param_u,
  param_d,
  param
)
}
\arguments{
\item{graph}{an object of class igraph. Can be both directed or undirected.}

\item{field_B}{the 'graph' edge attribute to be used to calculate the distance. Default is \code{"length"}.}

\item{dir_distance_type}{how directionality in B_ij calculations is dealt with:
\code{"symmetric"} (i.e. undirected graph) or \code{"asymmetric"} (i.e. directed graph). See details.}

\item{disp_type}{the formula used to calculate the probabilities in the B_ij matrix.
Use \code{"exponential"} for exponential decay, or \code{"threshold"} for setting a distance threshold.}

\item{param_u}{the upstream dispersal parameter. Must be a numeric value. Only used if \code{dir_distance_type = "asymmetric"}. See details.}

\item{param_d}{the downstream dispersal parameter. Must be a numeric value. Only used if \code{dir_distance_type = "asymmetric"}. See details.}

\item{param}{the dispersal parameter. Must be a numeric value. Only used if \code{dir_distance_type = "symmetric"}. See details.}
}
\value{
a matrix in data.frame format whose most relevant columns are 'from', 'to', and 'B_ij'.
For diagnosing purposes, also the distances are reported in the columns
'n' (undirected distance), 'u'(upstream distance), and 'd'(downstream distance).
}
\description{
Calculates B_ij: the coincidence probability due to dispersal
}
\details{
\code{dir_distance_type = "symmetric"} is to be used when the directionality of the river network is not relevant.
The distance between reaches midpoints is calculated for each couple of reaches.
\code{dir_distance_type = "asymmetric"} is to be used when the directionality is relevant.
The distance between reaches midpoints is calculated for each couple of reaches and splitted
between 'upstream travelled' distance and 'downstream travelled' distance

The 'param_u', 'param_d', and 'param' values are interpreted differently based on the formula used to relate distance and probability.
When \code{disp_type ="exponential"}, those values are used as the base of the exponential dispersal kernel: B_ij = param^{d_ij}.
When \code{disp_type ="threshold"}, those values are used to define the maximum dispersal length: B_ij = ifelse(d_ij < param, 1, 0).
}
