% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/c_ij_fun.R
\name{c_ij_fun}
\alias{c_ij_fun}
\title{Calculates c_ij: the coincidence probability due to the presence of barriers}
\usage{
c_ij_fun(
  graph,
  dir_fragmentation_type = "symmetric",
  pass_confluence = 1,
  pass_u = "pass_u",
  pass_d = "pass_d"
)
}
\arguments{
\item{graph}{an object of class igraph. Can be both directed or undirected.}

\item{dir_fragmentation_type}{how directionality in c_ij calculations is dealt with:
\code{"symmetric"} (i.e. undirected graph) or \code{"asymmetric"} (i.e. directed graph). See details.}

\item{pass_confluence}{a value in the range [0,1] that defines the passability of confluences (default is 1).}

\item{pass_u}{the 'graph' edge attribute to be used as upstream passability. Default is "pass_u".}

\item{pass_d}{the 'graph' edge attribute to be used as downstream passability. Default is "pass_d".}
}
\value{
a matrix in data.frame format whose columns are 'from', 'to', and 'c_ij'.
}
\description{
Calculates c_ij: the coincidence probability due to the presence of barriers
}
\details{
\code{dir_fragmentation_type = "symmetric"} is to be used when the directionality of the river network is not relevant.
The equivalent passability for each barrier is calculated as the product of upstream and downstream passabilities.
\code{dir_fragmentation_type = "asymmetric"} is to be used when the directionality is relevant.
The equivalent passability of each barrier is calculated as a function of the path connecting each couple of reaches
and depends on the direction of the path. Check the package vignette for more details.
}
