% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/index_calculation.R
\name{index_calculation}
\alias{index_calculation}
\title{Reach- and Catchment-scale indices of connectivity}
\usage{
index_calculation(
  graph,
  weight = "length",
  nodes_id = "name",
  index_type = "full",
  index_mode = "to",
  c_ij_flag = TRUE,
  B_ij_flag = TRUE,
  dir_fragmentation_type = "symmetric",
  pass_confluence = 1,
  pass_u = "pass_u",
  pass_d = "pass_d",
  field_B = "length",
  dir_distance_type = "symmetric",
  disp_type = "exponential",
  param_u,
  param_d,
  param
)
}
\arguments{
\item{graph}{an object of class igraph. Can be both directed or undirected.}

\item{weight}{graph vertex attribute used to assign weights to the reaches (nodes/vertices). Should not be also an edge attribute.
Default is \code{"length"}.}

\item{nodes_id}{graph vertex attribute used to univoquely label reaches (nodes/vertices). Should not be also an edge attribute.
Default is \code{"name"}. The graph attribute must be a character vector.
Used to label the results when \code{index_type = "reach"}}

\item{index_type}{indicates if the index should be calculated for the whole catchment (\code{index_type = "full"}),
for each reach (\code{index_type = "reach"}), or for each barrier (\code{index_type = "sum"})}

\item{index_mode}{indicates if reach index should be calculated based on inbound links ("to") or outbound links ("from").
Only active when \code{index_type = "reach"}.}

\item{c_ij_flag}{include in the calculation the barriers.}

\item{B_ij_flag}{include in the calculation the dispersal/movement among reaches.}

\item{dir_fragmentation_type}{how directionality in c_ij calculations is dealt with:
\code{"symmetric"} (i.e. undirected graph) or \code{"asymmetric"} (i.e. directed graph). See details below.}

\item{pass_confluence}{a value in the range [0,1] that defines the passability of confluences (default is 1).}

\item{pass_u}{the 'graph' edge attribute to be used as upstream passability. Default is "pass_u".}

\item{pass_d}{the 'graph' edge attribute to be used as downstream passability. Default is "pass_d".}

\item{field_B}{the 'graph' vertex attribute to be used to calculate the distance. Should not be also an edge attribute.
Default is \code{"length"}.}

\item{dir_distance_type}{how directionality in B_ij calculations is dealt with:
\code{"symmetric"} (i.e. undirected graph) or \code{"asymmetric"} (i.e. directed graph). See details below.}

\item{disp_type}{the formula used to calculate the probabilities in the B_ij matrix.
Use \code{"exponential"} for exponential decay, or \code{"threshold"} for setting a distance threshold.}

\item{param_u}{upstream dispersal parameter. Must be a numeric value.
Only used if \code{dir_distance_type = "asymmetric"}. See details below.}

\item{param_d}{downstream dispersal parameter.
Must be a numeric value. Only used if \code{dir_distance_type = "asymmetric"}. See below for details.}

\item{param}{dispersal parameter. Must be a numeric value.
Only used if \code{dir_distance_type = "symmetric"}. See details below.}
}
\value{
If \code{index_type = "full"}, returns a numeric value with the index value (column 'index').
if \code{index_type = c("reach", "sum")}, returns a data frame with the index value (column 'index') for each reach
(the field specified in 'nodes_id' is used for reach identification in the data frame).
In both cases, both numerator and denominator used in the index calculations are reported in the columns 'num' and 'den'.
}
\description{
Reach- and Catchment-scale indices of connectivity
}
\details{
Setting \code{c_ij_flag = FALSE} removes from the calculations the effect of barriers, i.e. the c_{ij} contribution
is not used in the calculation of the index.
Setting \code{B_ij_flag = FALSE} removes from the calculations the effect of movement/dispersal,
i.e. the B_{ij} contribution is not used in the calculation of the index.
Note that it is not possible to set both \code{c_ij_flag = FALSE} and \code{B_ij_flag = FALSE}.

The setting \code{dir_distance_type = "symmetric"} is to be used when the directionality of the river network is not relevant.
The distance between reaches midpoints is calculated for each couple of reaches.
The setting \code{dir_distance_type = "asymmetric"} is to be used when the directionality is relevant.
The distance between reaches midpoints is calculated for each couple of reaches and splitted
between 'upstream travelled' distance and 'downstream travelled' distance

The 'param_u', 'param_d', and 'param' values are interpreted differently based on the formula used to relate distance and probability.
When \code{disp_type ="exponential"}, those values are used as the base of the exponential dispersal kernel: B_ij = param^{d_ij}.
When \code{disp_type ="threshold"}, those values are used to define the maximum dispersal length: B_ij = ifelse(d_ij < param, 1, 0).
}
