% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/set_B_directionality.R
\name{set_B_directionality}
\alias{set_B_directionality}
\title{Calculates average upstream/downstream or combined distance between each couple of reaches}
\usage{
set_B_directionality(
  graph,
  dir_distance_type = "symmetric",
  field_B = "length"
)
}
\arguments{
\item{graph}{an object of class igraph. Can be both directed or undirected.}

\item{dir_distance_type}{how directionality in B_ij calculations is dealt with:
\code{"symmetric"} (i.e. undirected graph) or \code{"asymmetric"} (i.e. directed graph). See details.}

\item{field_B}{the 'graph' edge attribute to be used to calculate the distance. Default is \code{"length"}.}
}
\value{
an object of class 'igraph' with the upstream/downstream/total distances stored as edges attributes.
}
\description{
Calculates average upstream/downstream or combined distance between each couple of reaches
}
\keyword{internal}
