% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/t_index_calculation.R
\name{t_index_calculation}
\alias{t_index_calculation}
\title{Calculates time-dependent index when nodes weights or barriers passability are changing}
\usage{
t_index_calculation(
  graph,
  dams_metadata,
  id_dam = "id_dam",
  year = "year",
  pass_u = "pass_u",
  pass_d = "pass_d",
  weights_metadata,
  weight = "length",
  nodes_id = "name",
  parallel = TRUE,
  ncores,
  index_type = "full",
  index_mode = "from",
  ...,
  param_u,
  param_d,
  param
)
}
\arguments{
\item{graph}{an object of class igraph. Can be both directed or undirected.}

\item{dams_metadata}{data.frame that must contain a column having the same name as the 'id_dam' attribute of the graph,
and two columns with the corresponding upstream and downstream improved passabilities (see pass_u and pass_d), and a column with the year
passability was changed. This data frame can be obtained from easily-formatted data with the function \code{t_passability_sequencer}.}

\item{id_dam}{graph edges numeric attribute used to label dams. Default is \code{"id_dam"}. It should be present in the 'dams metadata' input as well.}

\item{year}{field of the 'dams metadata' where temporal information on the changes in passabiity is stored.}

\item{pass_u}{field of the 'dams metadata' where temporal-dependent upstream passabiity is stored.}

\item{pass_d}{field of the 'dams metadata' where temporal-dependent downstream passabiity is stored.}

\item{weights_metadata}{data.frame that must contain a column having the same name as the 'nodes_id' attribute of the graph,
a column with he corresponding weight information (see 'weight' parameter), and a column with the year
weight was changed. This data frame can be obtained from easily-formatted data with the function \code{t_weight_sequencer}.}

\item{weight}{param weight graph vertex attribute used to assign weights to the reaches (nodes). Default is \code{"length"}.}

\item{nodes_id}{graph vertex attribute used to univoquely label reaches (nodes). Default is \code{"name"}.}

\item{parallel}{logical value to flag if parallel option is to be used.}

\item{ncores}{define how many cores are used in parallel processing. Active only when \code{parallel = TRUE}}

\item{index_type}{indicates if the index should be calculated for the whole catchment (\code{index_type = "full"}),
or for each reach (\code{index_type = "reach"})}

\item{index_mode}{indicates if reach index should be calculated based on inbound links ("to") or outbound links ("from").
Only active when \code{index_type = "reach"}.}

\item{...}{other arguments passed to the function index_calculation}

\item{param_u}{upstream dispersal parameter. Must be a numeric value. Only used if \code{dir_distance_type = "asymmetric"}. See details.}

\item{param_d}{downstream dispersal parameter. Must be a numeric value. Only used if \code{dir_distance_type = "asymmetric"}. See details.}

\item{param}{dispersal parameter. Must be a numeric value. Only used if \code{dir_distance_type = "symmetric"}. See details.}
}
\value{
a data.frame with a 'year' field and related connectity index.
If \code{index_type = "reach"}, the data.frame is organized by 'year' and 'name'.
}
\description{
Calculates time-dependent index when nodes weights or barriers passability are changing
}
